+++ src/components/header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Script } from '../types';
+import { Script } from 'kit-bridge/cjs/type';
 
 interface HeaderProps {
   script: Script;
@@ -8,7 +8,7 @@ interface HeaderProps {
 
 export default function Header({ script, pid }: HeaderProps) {
   return (
-    <div className="text-xxs uppercase font-mono justify-between pt-3 px-4 grid grid-cols-5">
+    <div className="text-xxs uppercase font-mono font-bold justify-between pt-3 px-4 grid grid-cols-5">
       <span className="dark:text-primary-light text-primary-dark col-span-3">
         {script?.description || ''}
       </span>