+++ src/shortcuts.ts
@@ -1,10 +1,14 @@
 import { globalShortcut } from 'electron';
 import log from 'electron-log';
 import { readFile } from 'fs/promises';
+import { Script } from 'kit-bridge/cjs/type';
+import {
+  mainScriptPath,
+  shortcutsPath,
+  shortcutNormalizer,
+} from 'kit-bridge/cjs/util';
 import { runPromptProcess } from './kit';
-import { mainScriptPath, shortcutsPath, shortcutNormalizer } from './helpers';
 import { emitter, KitEvent } from './events';
-import { Script } from './types';
 
 const registerShortcut = (shortcut: string, filePath: string) => {
   const success = globalShortcut.register(shortcut, async () => {
@@ -76,6 +80,7 @@ export const updateMainShortcut = async (filePath: string) => {
       }
 
       const ret = globalShortcut.register(shortcut, async () => {
+        log.info(`🏚 main shortcut`);
         await runPromptProcess(mainScriptPath);
       });