+++ src/show.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/prefer-default-export */
 import { app, BrowserWindow, screen } from 'electron';
 import { writeFile, mkdir } from 'fs/promises';
+import { kenvPath, isDir } from 'kit-bridge/cjs/util';
 import { getAssetPath } from './assets';
-import { kenvPath, isDir } from './helpers';
 
 const page = (body: string) => {
   return `<!DOCTYPE html>
@@ -78,11 +78,11 @@ export const show = async (
     }
   });
 
-  const showParentDir = isDir(kenvPath('tmp'))
+  const showParentDir = (await isDir(kenvPath('tmp')))
     ? kenvPath('tmp', name)
     : app.getPath('appData');
 
-  if (!isDir(showParentDir)) {
+  if (!(await isDir(showParentDir))) {
     await mkdir(showParentDir, { recursive: true });
   }