+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+# [1.3.0-alpha.70](https://github.com/johnlindquist/kitapp/compare/v1.3.0-alpha.69...v1.3.0-alpha.70) (2021-07-08)
+
+
+### Features
+
+* **app:** extracting kit-bridge ([6df6372](https://github.com/johnlindquist/kitapp/commit/6df6372d16c1ee59d14666dd003f2afcd88b0570))
+
+# [1.3.0-alpha.69](https://github.com/johnlindquist/kitapp/compare/v1.3.0-alpha.68...v1.3.0-alpha.69) (2021-07-08)
+
+
+### Features
+
+* **app:** extracting kit-bridge ([904160e](https://github.com/johnlindquist/kitapp/commit/904160ef3374217cb47c4fc934f9f9feaef6e99c))
+
+# [1.3.0-alpha.68](https://github.com/johnlindquist/kitapp/compare/v1.3.0-alpha.67...v1.3.0-alpha.68) (2021-07-08)
+
+
+### Features
+
+* **app:** extracting kit-bridge ([453936e](https://github.com/johnlindquist/kitapp/commit/453936efea99774f8e6006f78f15fade30bac539))
+* **app:** extracting kit-bridge ([43672ac](https://github.com/johnlindquist/kitapp/commit/43672ac1d2bb165175a8dca581ddbf0d4e16eb9a))
+
+# [1.3.0-alpha.67](https://github.com/johnlindquist/kitapp/compare/v1.3.0-alpha.66...v1.3.0-alpha.67) (2021-07-08)
+
+
+### Features
+
+* **app:** extracting kit-bridge ([051a51a](https://github.com/johnlindquist/kitapp/commit/051a51add580b89d5613c252477f889993413d8b))
+* **app:** extracting kit-bridge ([3fb52e5](https://github.com/johnlindquist/kitapp/commit/3fb52e5d89457d29a602d486e0b800e89cc33feb))
+* **app:** extracting kit-bridge ([aae0849](https://github.com/johnlindquist/kitapp/commit/aae0849ab0951113c245c47d2d2b3b8bde8dfc03))
+* **app:** extracting kit-bridge ([6b0cead](https://github.com/johnlindquist/kitapp/commit/6b0cead8cf791fbae50d6fbe1c46c7d293334db4))
+* **app:** extracting kit-bridge ([6872f50](https://github.com/johnlindquist/kitapp/commit/6872f5041d065f3df4efb55327497f9240e77100))
+* **app:** extracting kit-bridge ([70348c1](https://github.com/johnlindquist/kitapp/commit/70348c12932ec1b3673ddab8018d4efa6e8965b8))
+* **app:** extracting kit-bridge ([a185c3d](https://github.com/johnlindquist/kitapp/commit/a185c3d72b96ce265e1ba8adaa3ca046ee1c8f83))
+* **app:** extracting kit-bridge ([de3d24a](https://github.com/johnlindquist/kitapp/commit/de3d24aecfd6d20977348c486b86481c1617dfce))
+* **app:** extracting kit-bridge ([27d20f4](https://github.com/johnlindquist/kitapp/commit/27d20f45986e1e4b5dc57dcb3df5c7d983ab2969))
+
+# [1.3.0-alpha.66](https://github.com/johnlindquist/kitapp/compare/v1.3.0-alpha.65...v1.3.0-alpha.66) (2021-07-02)
+
+
+### Features
+
+* **prompt:** handle kenvs ([30c0723](https://github.com/johnlindquist/kitapp/commit/30c072372387864792bd3df47bfe7c8ba7e89795))
+
 # [1.3.0-beta.20](https://github.com/johnlindquist/kitapp/compare/v1.3.0-beta.19...v1.3.0-beta.20) (2021-06-26)